+++ includes/service-providers/class-newspack-newsletters-service-provider-controller.php
@@ -34,4 +34,19 @@ abstract class Newspack_Newsletters_Service_Provider_Controller extends \WP_REST
 	public function register_routes() {
 		// Currently empty. Add endpoints common to all the ESP Service Providers.
 	}
+
+	/**
+	 * Update user's default email addresses for test email
+	 * 
+	 * @param array $emails Email addresses.
+	 * @return bool Whether the value has been updated or not
+	 */
+	public function update_user_test_emails( $emails ) {
+		$user_id   = get_current_user_id();
+		$user_info = get_userdata( $user_id );
+		if ( 1 === count( $emails ) && $user_info->user_email === $emails[0] ) {
+			return false;
+		}
+		return (bool) update_user_meta( $user_id, 'newspack_nl_test_emails', $emails );
+	}
 }