+++ webpack.config.js
@@ -16,6 +16,8 @@ const editor = path.join( __dirname, 'src', 'editor' );
 const admin = path.join( __dirname, 'src', 'admin' );
 const adsEditor = path.join( __dirname, 'src', 'ads-admin', 'editor' );
 const branding = path.join( __dirname, 'src', 'branding' );
+const quickEdit = path.join( __dirname, 'src', 'quick-edit' );
+const blocks = path.join( __dirname, 'src', 'editor', 'blocks' );
 
 const webpackConfig = getBaseWebpackConfig(
 	{ WP: true },
@@ -25,6 +27,8 @@ const webpackConfig = getBaseWebpackConfig(
 			admin,
 			adsEditor,
 			branding,
+			quickEdit,
+			blocks,
 		},
 		'output-path': path.join( __dirname, 'dist' ),
 	}